requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.10.5
sqlalchemy==2.0.35
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2